requirements.txt
python-dotenv
gpt4all==0.2.3
streamlit
pandas